+++ docs/conf.py
@@ -5,7 +5,7 @@ import ietfparse
 
 
 project = 'ietfparse'
-copyright = '2014-2017, Dave Shawley'
+copyright = '2014-2020, Dave Shawley'
 version = ietfparse.version
 release = '.'.join(str(x) for x in ietfparse.version_info[:2])
 
@@ -21,7 +21,7 @@ source_suffix = '.rst'
 source_encoding = 'utf-8-sig'
 master_doc = 'index'
 pygments_style = 'sphinx'
-html_static_path = []
+html_static_path = ['.']
 exclude_patterns = []
 html_sidebars = {
     '**': ['about.html', 'navigation.html', 'searchbox.html'],